+++ sdc/rabbit/publisher.py
@@ -2,12 +2,10 @@ import logging
 
 import pika
 from pika.exceptions import NackError, UnroutableError
-from structlog import wrap_logger
 
 from sdc.rabbit.exceptions import PublishMessageError
 
 logger = logging.getLogger(__name__)
-logger = wrap_logger(logger)
 
 
 class QueuePublisher(object):
@@ -23,6 +21,7 @@ class QueuePublisher(object):
         :param logger: A reference to a logging.Logger instance
         :param urls: List of RabbitMQ cluster URLs.
         :param queue: Queue name
+        :param confirm_delivery: Delivery confirmations toggle
         :param **kwargs: Custom key/value pairs passed to the arguments
             parameter of pika's channel.queue_declare method
 
@@ -35,6 +34,10 @@ class QueuePublisher(object):
         self._arguments = kwargs
         self._connection = None
         self._channel = None
+        self._confirm_delivery = False
+        if 'confirm_delivery' in kwargs:
+            self._confirm_delivery = True
+            self._arguments.pop('confirm_delivery', None)
 
     def _connect(self):
         """
@@ -52,12 +55,14 @@ class QueuePublisher(object):
                 self._channel.queue_declare(queue=self._queue,
                                             durable=self._durable_queue,
                                             arguments=self._arguments)
+                if self._confirm_delivery:
+                    self._channel.confirm_delivery()
+                    logger.info("Enabled delivery confirmation")
                 logger.debug("Connected to queue")
                 return True
 
-            except pika.exceptions.AMQPConnectionError as e:
-                logger.error("Unable to connect to queue",
-                             exception=repr(e))
+            except pika.exceptions.AMQPConnectionError:
+                logger.exception("Unable to connect to queue")
                 continue
 
         raise pika.exceptions.AMQPConnectionError
@@ -73,16 +78,18 @@ class QueuePublisher(object):
             self._connection.close()
             logger.debug("Disconnected from queue")
 
-        except Exception as e:
-            logger.error("Unable to close connection", exception=repr(e))
+        except Exception:
+            logger.exception("Unable to close connection")
 
-    def publish_message(self, message, content_type=None, headers=None):
+    def publish_message(self, message, content_type=None, headers=None, mandatory=False, immediate=False):
         """
         Publish a response message to a RabbitMQ queue.
 
         :param message: Response message
         :param content_type: Pika BasicProperties content_type value
         :param headers: Message header properties
+        :param mandatory: The mandatory flag
+        :param immediate: The immediate flag
 
         :returns: Boolean corresponding to the success of publishing
         :rtype: bool
@@ -96,14 +103,19 @@ class QueuePublisher(object):
             raise PublishMessageError
 
         try:
-            self._channel.basic_publish(exchange='',
-                                        routing_key=self._queue,
-                                        properties=pika.BasicProperties(
-                                            content_type=content_type,
-                                            headers=headers,
-                                            delivery_mode=2
-                                        ),
-                                        body=message)
+            result = self._channel.basic_publish(exchange='',
+                                                 routing_key=self._queue,
+                                                 mandatory=mandatory,
+                                                 immediate=immediate,
+                                                 properties=pika.BasicProperties(
+                                                     content_type=content_type,
+                                                     headers=headers,
+                                                     delivery_mode=2
+                                                 ),
+                                                 body=message)
+
+            logger.info('Published message to queue', queue=self._queue)
+            return result
         except NackError:
             # raised when a message published in publisher-acknowledgments mode
             # is returned via `Basic.Return` followed by `Basic.Ack`.
@@ -115,7 +127,5 @@ class QueuePublisher(object):
             logger.error("UnroutableError occured. Message not published.")
             raise PublishMessageError
         except Exception:
-            logger.error("Unknown exception occured. Message not published.")
+            logger.exception("Unknown exception occured. Message not published.")
             raise PublishMessageError
-        msg = 'Published message to {} queue'
-        logger.error(msg.format(self._queue))